Add files via upload
The Python URL Shortener project is a simple yet powerful tool that allows you to shorten long URLs into concise and easy-to-share links. With just a few lines of Python code, you can create your very own URL shortening service.

This project leverages the flexibility and versatility of the Python programming language to provide a seamless user experience. It takes a long URL as input and generates a unique short code that redirects users to the original URL when accessed.